[node.sh] moved blsloader folder to internal/blsgen
