+++ internal/blsgen/kms_test.go
@@ -1,4 +1,4 @@
-package blsloader
+package blsgen
 
 import (
 	"encoding/json"
