+++ internal/blsgen/utils_test.go
@@ -1,4 +1,4 @@
-package blsloader
+package blsgen
 
 import (
 	"fmt"
